requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
boto3==1.42.96
botocore==1.42.96
boxsdk<4
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
cloudpathlib==0.23.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.19.1
filetype==1.2.0
fire==0.7.1
fonttools==4.60.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.47
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
kagglehub==0.3.13
kiwisolver==1.4.7
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
onnx==1.19.1
onnx2torch==1.5.15
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==26.2
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==11.3.0
pillow-avif-plugin==1.5.5
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pybboxes==0.1.6
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
reverse_geocoder==1.5.1
rich==15.0.0
roboflow==1.2.16
s3transfer==0.16.1
sahi==0.11.36
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
speciesnet==5.0.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
traitlets==5.14.3
typer==0.23.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==1.26.20
wcwidth==0.6.0
Werkzeug==3.1.8
yolov5==7.0.11
zipp==3.23.1
